# public/wmcp-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Co-Achat Loyalty Card - Public Merchant API",
    "description": "Public API for discovering merchants with geolocation, loyalty programs, promotional offers, and Google Reviews integration. This API is freely accessible without authentication.",
    "version": "1.0.0",
    "contact": {
      "name": "Co-Achat Support",
      "url": "https://coachat-loyalty-card.test"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://coachat-loyalty-card.test/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Development server"
    },
    {
      "url": "http://coachat-backend-loyalty-card.test/api",
      "description": "Local environment"
    }
  ],
...
  "paths": {
    "/public/wmcp/merchants/map": {
      "get": {
        "tags": ["Public"],
        "summary": "Get claimed merchants map",
        "description": "Discover claimed merchants only with geolocation, loyalty programs, promotional offers, scheduled and active contests, and Google Reviews configuration. Unclaimed merchants are excluded.",
        "operationId": "getWmcpClaimedMerchantsMap",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "description": "Latitude of search center point",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "example": 48.8566
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "description": "Longitude of search center point",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "example": 2.3522
            }
          },
          {
            "name": "radius_km",
            "in": "query",
            "description": "Search radius in kilometers (default: 10)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "default": 10,
              "example": 5
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default: 20, max: 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100,
              "example": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "example": 0
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Text search on merchant name (case-insensitive)",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Boulangerie"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of merchants with their details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantsMapResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalid_lat_lon": {
                    "value": {
                      "error": "Invalid latitude/longitude",
                      "message": "latitude and longitude must be valid numbers"
                    }
                  },
                  "invalid_radius": {
                    "value": {
                      "error": "Invalid radius",
                      "message": "radius_km must be positive"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/wmcp/contests": {
      "get": {
        "tags": ["Public"],
        "summary": "List public contests with rewards",
        "description": "List scheduled and active contests (jeux concours) across claimed merchants, including the contest schedule and every reward (lot) to win. Useful for AI discovery and aggregators.",
        "operationId": "getWmcpPublicContests",
        "parameters": [
          {
            "name": "merchant_id",
            "in": "query",
            "description": "Restrict listing to a single merchant (UUID)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of contests (default 50, max 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (default 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of public contests with rewards",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicContestsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MerchantsMapResponse": {
        "type": "object",
        "description": "Response containing list of merchants for map view",
        "required": ["merchants", "total"],
        "properties": {
          "merchants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapMerchant"
            },
            "description": "List of merchants matching search criteria"
          },
          "total": {
            "type": "integer",
            "description": "Total number of merchants matching criteria (without pagination)",
            "example": 145
          }
        }
      },
      "MapMerchant": {
        "type": "object",
        "description": "Merchant data for map display",
        "required": [
          "id",
          "company_name",
          "latitude",
          "longitude",
          "has_active_content",
          "is_customer_linked",
          "loyalty_programs",
          "active_promotional_offers"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique merchant identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "company_name": {
            "type": "string",
            "description": "Merchant business name",
            "example": "Boulangerie Martin"
          },
          "address": {
            "type": ["string", "null"],
            "description": "Complete street address",
            "example": "123 rue de la Paix, 75000 Paris"
          },
          "postal_code": {
            "type": ["string", "null"],
            "description": "Postal code",
            "example": "75000"
          },
          "city": {
            "type": ["string", "null"],
            "description": "City name",
            "example": "Paris"
          },
          "phone": {
            "type": ["string", "null"],
            "description": "Merchant phone number",
            "example": "01 23 45 67 89"
          },
          "logo_url": {
            "type": ["string", "null"],
            "description": "Merchant logo as base64 data URL",
            "example": "data:image/png;base64,iVBORw0KGgoAAAANS..."
          },
          "latitude": {
            "type": "number",
            "format": "float",
            "description": "Geographic latitude",
            "example": 48.8566
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "description": "Geographic longitude",
            "example": 2.3522
          },
          "distance_km": {
            "type": ["number", "null"],
            "format": "float",
            "description": "Distance from search center in kilometers (null if no geolocation provided)",
            "example": 2.5
          },
          "subscriber_count": {
            "type": ["integer", "null"],
            "description": "Number of active subscribers (social proof)",
            "example": 42
          },
          "has_active_content": {
            "type": "boolean",
            "description": "Merchant has at least one active loyalty program or promotional offer",
            "example": true
          },
          "is_customer_linked": {
            "type": "boolean",
            "description": "Current user is linked to this merchant (requires customer authentication)",
            "example": false
          },
          "loyalty_programs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapLoyaltyProgram"
            },
            "description": "Active loyalty programs"
          },
          "active_promotional_offers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapActiveOffer"
            },
            "description": "Active promotional offers"
          },
          "active_contests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicContest"
            },
            "description": "Scheduled or active contests (jeux concours) with the list of rewards to win"
          },
          "google_review": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MapGoogleReview"
              },
              {
                "type": "null",
                "description": "Null if Google Review module is not fully configured"
              }
            ],
            "description": "Google Review module configuration (null if incomplete)"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code or type"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          }
        }
      },
      "PublicContestsResponse": {
        "type": "object",
        "description": "Response payload for the public contests listing endpoint",
        "required": ["contests", "total"],
        "properties": {
          "contests": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicContestWithMerchant" }
          },
          "total": {
            "type": "integer",
            "description": "Number of contests returned in this page"
          }
        }
      },
...
      "PublicContestWithMerchant": {
        "allOf": [
          { "$ref": "#/components/schemas/PublicContest" },
          {
            "type": "object",
            "required": ["merchant"],
            "properties": {
              "merchant": {
                "type": "object",
                "description": "Merchant organising the contest",
                "required": ["id", "company_name"],
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "company_name": { "type": "string" },
                  "city": { "type": ["string", "null"] },
                  "postal_code": { "type": ["string", "null"] },
                  "latitude": { "type": ["number", "null"], "format": "float" },
                  "longitude": { "type": ["number", "null"], "format": "float" },
                  "logo_url": { "type": ["string", "null"] }
                }
              }
            }
          }
        ]
      },
...
    }
  }
}